requirements.txt
annotated-types==0.6.0
anyio==4.3.0
atproto==0.0.46
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
idna==3.7
libipld==1.2.3
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
sniffio==1.3.1
typing_extensions==4.11.0
websockets==12.0
